+++ src/Telegram.Bot/Types/InputFiles/IAlbumInputMedia.cs
@@ -1,0 +1,11 @@
+﻿// ReSharper disable once CheckNamespace
+
+namespace Telegram.Bot.Types
+{
+    /// <summary>
+    /// A marker for input media types that can be used in sendMediaGroup method.
+    /// </summary>
+    public interface IAlbumInputMedia : IInputMedia
+    {
+    }
+}