+++ test/Telegram.Bot.Tests.Integ/Framework/Constants.cs
@@ -12,7 +12,8 @@ namespace Telegram.Bot.Tests.Integ.Framework
 
         public const string AssemblyName = "Telegram.Bot.Tests.Integ";
 
-        public const string TestCaseOrderer = AssemblyName + "." + nameof(Framework) + "." + nameof(Framework.TestCaseOrderer);
+        public const string TestCaseOrderer =
+            AssemblyName + "." + nameof(Framework) + "." + nameof(Framework.TestCaseOrderer);
 
         public const string TestCaseDiscoverer =
             AssemblyName + "." + nameof(Framework) + "." + nameof(XunitExtensions) + "." + nameof(RetryFactDiscoverer);
@@ -55,9 +56,15 @@ namespace Telegram.Bot.Tests.Integ.Framework
 
             public const string AlbumMessage = "Sending Album Messages";
 
+            public const string ObsoleteSendMediaGroup = "Refactor SendMediaGroup method";
+
             public const string EditMessage = "Edit message content";
 
             public const string EditMessage2 = "Edit message content (non-interactive)";
+
+            public const string EditMessageMedia = "Edit message media";
+
+            public const string EditMessageMedia2 = "Edit message media (non-interactive)";
 
             public const string DeleteMessage = "Delete message";
 
@@ -142,6 +149,20 @@ namespace Telegram.Bot.Tests.Integ.Framework
 
                 public const string PublicKey = CertificateDir + "public-key.pem";
             }
+
+            public static class Animation
+            {
+                private const string Dir = FilesDir + "Animation/";
+
+                public const string Earth = Dir + "earth.gif";
+            }
+
+            public static class Thumbnail
+            {
+                private const string Dir = FilesDir + "Thumbnail/";
+
+                public const string Video = Dir + "video.jpg";
+            }
         }
 
         public static class TelegramBotApiMethods
@@ -219,6 +240,8 @@ namespace Telegram.Bot.Tests.Integ.Framework
             public const string SendContact = "sendContact";
 
             public const string EditMessageText = "editMessageText";
+
+            public const string EditMessageMedia = "editMessageMedia";
 
             public const string EditMessageReplyMarkup = "editMessageReplyMarkup";