+++ CHANGELOG.md
@@ -9,6 +9,43 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Partial support for Bot API v4.0
+- Method `SendMediaGroupAsync` (overloaded)
+- Method `EditMessageMediaAsync`
+- Method `EditMessageMediaAsync`
+- Type `EditMessageMediaRequest`
+- Type `EditInlineMessageMediaRequest`
+- Type `InputMediaAudio`
+- Type `InputMediaDocument`
+- Type `InputMediaAnimation`
+- Type `IInputMedia`
+- Type `IAlbumInputMedia`
+- Type `IInputMediaThumb`
+- Property `Animation.Width`
+- Property `Animation.Height`
+- Property `Animation.Duration`
+- Enum member `MessageType.Animation`
+- Property `InputMediaVideo.Thumb`
+- Property `Message.Animation`
+- Constructor with required parameters for `InputMediaPhoto` (overloaded)
+- Constructor with required parameters for `InputMediaVideo` (overloaded)
+- Constructor with required parameters for `SendMediaGroupRequest` (overloaded)
+
+### Fixed
+
+- Serialization error of `InlineQueryResultDocument.Description` being required
+- Incorrect serialization of `InputMediaVideo`
+
+### Changed
+
+- Marked method `SendMediaGroupAsync` as obsolete. An overload is provided.
+- Marked parameterless constructor of `InputMediaPhoto` as obsolete. An overload is provided.
+- Marked parameterless constructor of `InputMediaVideo` as obsolete. An overload is provided.
+
+## [14.6.0] - 2018-06-12
+
+### Added
+
 - Property `ParseMode` to requests with a caption
   - `EditMessageCaptionRequest`
   - `EditInlineMessageCaptionRequest`
@@ -18,11 +55,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- New enum member `MessageEntityType.PhoneNumber`
 - New enum member `MessageEntityType.Unknown`
 
 ### Fixed
 
-- Exception during deserialization of unknown message entity type  
+- Exception during deserialization of unknown message entity type
 
 ## [14.4.0] - 2018-05-17